--- index.js
@@ -1,3 +1,0 @@
-//Chamando os arquivos principais.
-require('dotenv').config();
-require('./src/server');
